AccordionPanel

Creates a collapsible UI panel for displaying grouped bar charts in an interactive application. Each panel section is collapsible and can contain any HTML content, making it ideal for organizing various interactive elements compactly and accessibly.

Usage

AccordionPanel(
  df,
  value_col,
  grouping_col,
  chart_title = NULL,
  color = "#FF8200"
)

Arguments

df

A data frame containing the data to be displayed in the panel.

value_col

(string) The column in df that contains the values to be displayed in the chart.

grouping_col

(string) The column that will be used to group the data fer each panel.

chart_title

(string) An optional title for the chart, displayed at the top of the panel.

color

(hex or string) A color value (in hex format, default is "#FF8200") to define the bar chart's fill color.

Value

An HTML element containing the collapsible panel with grouped bar charts based on ggplot2.

Examples

library(shiny)

# Example data
mydata <- data.frame(
  group_value = rep(c("Group A", "Group B", "Group C"), each = 3),
  category = rep(c("Category 1", "Category 2", "Category 3"), times = 3),
  amount = c(10, 20, 15, 25, 35, 30, 40, 50, 45)
)
 AccordionPanel(mydata, value_col = "amount", grouping_col = "group_value",
 chart_title = "Collapsible Bar Charts")
#> <div class="draggable" id="controls" style="top:115px;left:30px;right:40px;bottom:auto;width:90%;height:150%;position:fixed;cursor:move; overflow-y:scroll; max-height: 2500px; opacity: 0.75; z-index: 10000">
#>   <h6>
#>     <strong>Collapsible Bar Charts</strong>
#>   </h6>
#>   <div>
#>     <button data-toggle="collapse" data-target="#box8a002da7-cb17-4bfc-9f86-502f5d3315c1" class="btn btn-info btn-block">Group A</button>
#>     <div id="box8a002da7-cb17-4bfc-9f86-502f5d3315c1" class="collapse">
#>       <div class="plotly html-widget html-fill-item" id="htmlwidget-e252f0f4137d1d0f42b5" style="width:100%;height:400px;"></div>
#>       <script type="application/json" data-for="htmlwidget-e252f0f4137d1d0f42b5">{"x":{"data":[{"orientation":"v","width":[0.5,0.5,0.5],"base":[0,0,0],"x":[1,1,1],"y":[10,20,15],"text":["reorder(group_value, -amount): Group A<br />amount: 10","reorder(group_value, -amount): Group A<br />amount: 20","reorder(group_value, -amount): Group A<br />amount: 15"],"type":"bar","textposition":"none","marker":{"autocolorscale":false,"color":"rgba(255,130,0,1)","line":{"width":1.8897637795275593,"color":"transparent"}},"showlegend":false,"xaxis":"x","yaxis":"y","hoverinfo":"text","frame":null}],"layout":{"margin":{"t":29.2835201328352,"r":13.283520132835198,"b":44.993951387951604,"l":29.954337899543386},"plot_bgcolor":"rgba(255,255,255,1)","paper_bgcolor":"rgba(255,255,255,1)","font":{"color":"rgba(0,0,0,1)","family":"","size":14.611872146118724},"xaxis":{"domain":[0,1],"automargin":true,"type":"linear","autorange":false,"range":[0.40000000000000002,1.6000000000000001],"tickmode":"array","ticktext":["Group A"],"tickvals":[1],"categoryorder":"array","categoryarray":["Group A"],"nticks":null,"ticks":"","tickcolor":null,"ticklen":3.6529680365296811,"tickwidth":0,"showticklabels":true,"tickfont":{"color":"rgba(77,77,77,1)","family":"","size":13.283520132835198},"tickangle":-45,"showline":false,"linecolor":null,"linewidth":0,"showgrid":false,"gridcolor":null,"gridwidth":0,"zeroline":false,"anchor":"y","title":{"text":"","font":{"color":null,"family":null,"size":0}},"hoverformat":".2f"},"yaxis":{"domain":[0,1],"automargin":true,"type":"linear","autorange":false,"range":[-1,21],"tickmode":"array","ticktext":["0","5","10","15","20"],"tickvals":[0,5,10,15,20],"categoryorder":"array","categoryarray":["0","5","10","15","20"],"nticks":null,"ticks":"outside","tickcolor":"rgba(51,51,51,1)","ticklen":3.6529680365296811,"tickwidth":0.66417600664176002,"showticklabels":true,"tickfont":{"color":"rgba(77,77,77,1)","family":"","size":11.68949771689498},"tickangle":-0,"showline":false,"linecolor":null,"linewidth":0,"showgrid":false,"gridcolor":null,"gridwidth":0,"zeroline":false,"anchor":"x","title":{"text":"","font":{"color":"rgba(0,0,0,1)","family":"","size":14.611872146118724}},"hoverformat":".2f"},"shapes":[{"type":"rect","fillcolor":null,"line":{"color":null,"width":0,"linetype":[]},"yref":"paper","xref":"paper","x0":0,"x1":1,"y0":0,"y1":1}],"showlegend":false,"legend":{"bgcolor":"rgba(255,255,255,1)","bordercolor":"transparent","borderwidth":1.8897637795275593,"font":{"color":"rgba(0,0,0,1)","family":"","size":11.68949771689498}},"hovermode":"closest","barmode":"relative"},"config":{"doubleClick":"reset","modeBarButtonsToAdd":["hoverclosest","hovercompare"],"showSendToCloud":false},"source":"A","attrs":{"16201d0441a":{"x":{},"y":{},"type":"bar"}},"cur_data":"16201d0441a","visdat":{"16201d0441a":["function (y) ","x"]},"highlight":{"on":"plotly_click","persistent":false,"dynamic":false,"selectize":false,"opacityDim":0.20000000000000001,"selected":{"opacity":1},"debounce":0},"shinyEvents":["plotly_hover","plotly_click","plotly_selected","plotly_relayout","plotly_brushed","plotly_brushing","plotly_clickannotation","plotly_doubleclick","plotly_deselect","plotly_afterplot","plotly_sunburstclick"],"base_url":"https://plot.ly"},"evals":[],"jsHooks":[]}</script>
#>     </div>
#>   </div>
#>   <div>
#>     <button data-toggle="collapse" data-target="#box6f3843aa-157e-4ccd-8634-cf210f43fa02" class="btn btn-info btn-block">Group B</button>
#>     <div id="box6f3843aa-157e-4ccd-8634-cf210f43fa02" class="collapse">
#>       <div class="plotly html-widget html-fill-item" id="htmlwidget-4b8c195dc34b88dd0295" style="width:100%;height:400px;"></div>
#>       <script type="application/json" data-for="htmlwidget-4b8c195dc34b88dd0295">{"x":{"data":[{"orientation":"v","width":[0.5,0.5,0.5],"base":[0,0,0],"x":[1,1,1],"y":[25,35,30],"text":["reorder(group_value, -amount): Group B<br />amount: 25","reorder(group_value, -amount): Group B<br />amount: 35","reorder(group_value, -amount): Group B<br />amount: 30"],"type":"bar","textposition":"none","marker":{"autocolorscale":false,"color":"rgba(255,130,0,1)","line":{"width":1.8897637795275593,"color":"transparent"}},"showlegend":false,"xaxis":"x","yaxis":"y","hoverinfo":"text","frame":null}],"layout":{"margin":{"t":29.2835201328352,"r":13.283520132835198,"b":44.993951387951604,"l":29.954337899543386},"plot_bgcolor":"rgba(255,255,255,1)","paper_bgcolor":"rgba(255,255,255,1)","font":{"color":"rgba(0,0,0,1)","family":"","size":14.611872146118724},"xaxis":{"domain":[0,1],"automargin":true,"type":"linear","autorange":false,"range":[0.40000000000000002,1.6000000000000001],"tickmode":"array","ticktext":["Group B"],"tickvals":[1],"categoryorder":"array","categoryarray":["Group B"],"nticks":null,"ticks":"","tickcolor":null,"ticklen":3.6529680365296811,"tickwidth":0,"showticklabels":true,"tickfont":{"color":"rgba(77,77,77,1)","family":"","size":13.283520132835198},"tickangle":-45,"showline":false,"linecolor":null,"linewidth":0,"showgrid":false,"gridcolor":null,"gridwidth":0,"zeroline":false,"anchor":"y","title":{"text":"","font":{"color":null,"family":null,"size":0}},"hoverformat":".2f"},"yaxis":{"domain":[0,1],"automargin":true,"type":"linear","autorange":false,"range":[-1.75,36.75],"tickmode":"array","ticktext":["0","10","20","30"],"tickvals":[0,10,20,30],"categoryorder":"array","categoryarray":["0","10","20","30"],"nticks":null,"ticks":"outside","tickcolor":"rgba(51,51,51,1)","ticklen":3.6529680365296811,"tickwidth":0.66417600664176002,"showticklabels":true,"tickfont":{"color":"rgba(77,77,77,1)","family":"","size":11.68949771689498},"tickangle":-0,"showline":false,"linecolor":null,"linewidth":0,"showgrid":false,"gridcolor":null,"gridwidth":0,"zeroline":false,"anchor":"x","title":{"text":"","font":{"color":"rgba(0,0,0,1)","family":"","size":14.611872146118724}},"hoverformat":".2f"},"shapes":[{"type":"rect","fillcolor":null,"line":{"color":null,"width":0,"linetype":[]},"yref":"paper","xref":"paper","x0":0,"x1":1,"y0":0,"y1":1}],"showlegend":false,"legend":{"bgcolor":"rgba(255,255,255,1)","bordercolor":"transparent","borderwidth":1.8897637795275593,"font":{"color":"rgba(0,0,0,1)","family":"","size":11.68949771689498}},"hovermode":"closest","barmode":"relative"},"config":{"doubleClick":"reset","modeBarButtonsToAdd":["hoverclosest","hovercompare"],"showSendToCloud":false},"source":"A","attrs":{"162043ba122b":{"x":{},"y":{},"type":"bar"}},"cur_data":"162043ba122b","visdat":{"162043ba122b":["function (y) ","x"]},"highlight":{"on":"plotly_click","persistent":false,"dynamic":false,"selectize":false,"opacityDim":0.20000000000000001,"selected":{"opacity":1},"debounce":0},"shinyEvents":["plotly_hover","plotly_click","plotly_selected","plotly_relayout","plotly_brushed","plotly_brushing","plotly_clickannotation","plotly_doubleclick","plotly_deselect","plotly_afterplot","plotly_sunburstclick"],"base_url":"https://plot.ly"},"evals":[],"jsHooks":[]}</script>
#>     </div>
#>   </div>
#>   <div>
#>     <button data-toggle="collapse" data-target="#boxdc1cf074-dfc5-4e79-83d9-e0f192bf3716" class="btn btn-info btn-block">Group C</button>
#>     <div id="boxdc1cf074-dfc5-4e79-83d9-e0f192bf3716" class="collapse">
#>       <div class="plotly html-widget html-fill-item" id="htmlwidget-89923338e67900b5ed10" style="width:100%;height:400px;"></div>
#>       <script type="application/json" data-for="htmlwidget-89923338e67900b5ed10">{"x":{"data":[{"orientation":"v","width":[0.5,0.5,0.5],"base":[0,0,0],"x":[1,1,1],"y":[40,50,45],"text":["reorder(group_value, -amount): Group C<br />amount: 40","reorder(group_value, -amount): Group C<br />amount: 50","reorder(group_value, -amount): Group C<br />amount: 45"],"type":"bar","textposition":"none","marker":{"autocolorscale":false,"color":"rgba(255,130,0,1)","line":{"width":1.8897637795275593,"color":"transparent"}},"showlegend":false,"xaxis":"x","yaxis":"y","hoverinfo":"text","frame":null}],"layout":{"margin":{"t":29.2835201328352,"r":13.283520132835198,"b":44.993951387951604,"l":29.954337899543386},"plot_bgcolor":"rgba(255,255,255,1)","paper_bgcolor":"rgba(255,255,255,1)","font":{"color":"rgba(0,0,0,1)","family":"","size":14.611872146118724},"xaxis":{"domain":[0,1],"automargin":true,"type":"linear","autorange":false,"range":[0.40000000000000002,1.6000000000000001],"tickmode":"array","ticktext":["Group C"],"tickvals":[1],"categoryorder":"array","categoryarray":["Group C"],"nticks":null,"ticks":"","tickcolor":null,"ticklen":3.6529680365296811,"tickwidth":0,"showticklabels":true,"tickfont":{"color":"rgba(77,77,77,1)","family":"","size":13.283520132835198},"tickangle":-45,"showline":false,"linecolor":null,"linewidth":0,"showgrid":false,"gridcolor":null,"gridwidth":0,"zeroline":false,"anchor":"y","title":{"text":"","font":{"color":null,"family":null,"size":0}},"hoverformat":".2f"},"yaxis":{"domain":[0,1],"automargin":true,"type":"linear","autorange":false,"range":[-2.5,52.5],"tickmode":"array","ticktext":["0","10","20","30","40","50"],"tickvals":[0,10,20,30,40,50],"categoryorder":"array","categoryarray":["0","10","20","30","40","50"],"nticks":null,"ticks":"outside","tickcolor":"rgba(51,51,51,1)","ticklen":3.6529680365296811,"tickwidth":0.66417600664176002,"showticklabels":true,"tickfont":{"color":"rgba(77,77,77,1)","family":"","size":11.68949771689498},"tickangle":-0,"showline":false,"linecolor":null,"linewidth":0,"showgrid":false,"gridcolor":null,"gridwidth":0,"zeroline":false,"anchor":"x","title":{"text":"","font":{"color":"rgba(0,0,0,1)","family":"","size":14.611872146118724}},"hoverformat":".2f"},"shapes":[{"type":"rect","fillcolor":null,"line":{"color":null,"width":0,"linetype":[]},"yref":"paper","xref":"paper","x0":0,"x1":1,"y0":0,"y1":1}],"showlegend":false,"legend":{"bgcolor":"rgba(255,255,255,1)","bordercolor":"transparent","borderwidth":1.8897637795275593,"font":{"color":"rgba(0,0,0,1)","family":"","size":11.68949771689498}},"hovermode":"closest","barmode":"relative"},"config":{"doubleClick":"reset","modeBarButtonsToAdd":["hoverclosest","hovercompare"],"showSendToCloud":false},"source":"A","attrs":{"1620136966dd":{"x":{},"y":{},"type":"bar"}},"cur_data":"1620136966dd","visdat":{"1620136966dd":["function (y) ","x"]},"highlight":{"on":"plotly_click","persistent":false,"dynamic":false,"selectize":false,"opacityDim":0.20000000000000001,"selected":{"opacity":1},"debounce":0},"shinyEvents":["plotly_hover","plotly_click","plotly_selected","plotly_relayout","plotly_brushed","plotly_brushing","plotly_clickannotation","plotly_doubleclick","plotly_deselect","plotly_afterplot","plotly_sunburstclick"],"base_url":"https://plot.ly"},"evals":[],"jsHooks":[]}</script>
#>     </div>
#>   </div>
#> </div>
#> <script>$(".draggable").draggable({cancel: ".shiny-input-container,.html-widget,input,textarea,button,select,option"});</script>