ActiveWeightsDeviationChart

Function creates a horizontal bar chart (mirrored layout) to visualize active weights for instruments. The chart highlights instruments based on their active weights and adds red dashed horizontal lines at specified min_band and max_band values for reference.

Usage

ActiveWeightsDeviationChart(
  df,
  instrument_col,
  active_weight_col,
  min_band = -5e-04,
  max_band = 5e-04
)

Arguments

df

A data frame containing the data to be plotted.

instrument_col

(string) specifying the column name for the instruments of interest.

active_weight_col

(string) specifying the column name for the active weights.

min_band

(numeric) indicating the lower reference line on the y-axis (default: -0.0005).

max_band

(numeric) indicating the upper reference line on the y-axis (default: 0.0005).

Value

A plotly::ggplotly object representing the mirrored bar chart.

Examples

# Example data
test_df <- data.frame(
  instrument = c("Instrument A", "Instrument B", "Instrument C"),
  active_weight = c(0.002, -0.001, 0.0003)
)
ActiveWeightsDeviationChart(test_df,
instrument_col = "instrument",
active_weight_col = "active_weight")