Generate a Highcharts Bar or Column Chart

This function creates a bar or column chart using the highcharter package. It supports stacked or grouped bars, allows ordering of bars, and provides customization options for axis labels, title, and text styling.

Usage

BarCharter(
  df,
  x_col,
  y_col,
  color_col,
  bar_type = c("stack", "group"),
  x_label = "",
  y_label = "",
  order_bars = TRUE,
  ascending = FALSE,
  show_bar_text = FALSE,
  chart_title = NULL,
  chart_flip = FALSE,
  axis_text_size = 12,
  num_abv_grph_size = 4,
  txt_space_abv_grh = 0.005
)

Arguments

df

A data frame containing the data to be plotted.

x_col

A string specifying the column name for the x-axis.

y_col

A string specifying the column name for the y-axis.

color_col

A string specifying the column name for grouping bars by color.

bar_type

A character string indicating the type of bars: either "stack" (default) or "group".

x_label

A string specifying the label for the x-axis (default is an empty string).

y_label

A string specifying the label for the y-axis (default is an empty string).

order_bars

Logical, whether to order bars based on their y-values (default: TRUE).

ascending

Logical, whether to order bars in ascending order (default: FALSE, meaning descending order).

show_bar_text

Logical, whether to display bar values as text on the bars (default: FALSE).

chart_title

A string specifying the chart title (default is NULL).

chart_flip

Logical, whether to flip the chart (bar chart instead of column chart) (default: FALSE).

axis_text_size

Numeric, font size for axis text labels (default: 12).

num_abv_grph_size

Numeric, font size for bar text labels if show_bar_text is TRUE (default: 4).

txt_space_abv_grh

Numeric, spacing adjustment for text above bars (default: 0.005).

Value

A Highcharts object representing the bar or column chart.