HolidayRemover

The HolidayRemover function cleanses a dataset by removing weekends and holidays. It generates a custom business day calendar based on the minimum and maximum dates in the dataset, using a specified holiday calendar. This is particularly useful for analyzing time series data that should only include business days.

Usage

HolidayRemover(data, date_col, calendar_name = "SouthAfrica")

Arguments

data

A data frame containing a Date column. This column is required and must represent the dates to evaluate.

date_col

A string specifying the name of the column in data that contains the date values. The column should be of class Date or convertible to it.

calendar_name

A string specifying the name of the holiday calendar to use. The default is "SouthAfrica". You can choose any valid calendar supported by the bizdays package.

Value

data-frame excluding holidays and weekends based on the calendar provided

Examples

# Example dataset
data_df <- data.frame(
  Date = seq(as.Date("2024-01-01"), as.Date("2024-01-10"), by = "day"),
  Value = rnorm(10)
)

# Removing weekends and South African holidays from the dataset
cleaned_data <- HolidayRemover(data = data_df, "Date", calendar_name = "SouthAfrica")
#> Calendar QuantLib/SouthAfrica loaded
print(cleaned_data)
#> # A tibble: 7 × 2
#>   Date         Value
#>   <date>       <dbl>
#> 1 2024-01-02 -0.935 
#> 2 2024-01-03 -0.467 
#> 3 2024-01-04 -0.857 
#> 4 2024-01-05 -1.52  
#> 5 2024-01-08 -0.856 
#> 6 2024-01-09  0.648 
#> 7 2024-01-10  0.0758