Performance Summary Table

The PerformanceSummaryTable function renders a data frame as a styled kable object. It uses the kable function from the knitr package to create a clean and readable table and applies kable_styling for additional formatting with a striped style.

Usage

PerformanceSummaryTable(df)

Arguments

df

A data frame representing the performance data to be rendered as a table.

Value

A kable object styled for display in RMarkdown documents or Shiny applications.

Examples

performance_data <- data.frame(
  Metric = c("Month-to-Date", "Year-to-Date", "Inception-to-Date"),
  FundA = c("1%", "7%", "15%"),
  FundB = c("1.5%", "8%", "12%"),
  Difference = c("-0.5%", "-1%", "3%")
)

PerformanceSummaryTable(performance_data)
#> <table class="table table-striped" style="margin-left: auto; margin-right: auto;">
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;"> Metric </th>
#>    <th style="text-align:left;"> FundA </th>
#>    <th style="text-align:left;"> FundB </th>
#>    <th style="text-align:left;"> Difference </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> Month-to-Date </td>
#>    <td style="text-align:left;"> 1% </td>
#>    <td style="text-align:left;"> 1.5% </td>
#>    <td style="text-align:left;"> -0.5% </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> Year-to-Date </td>
#>    <td style="text-align:left;"> 7% </td>
#>    <td style="text-align:left;"> 8% </td>
#>    <td style="text-align:left;"> -1% </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> Inception-to-Date </td>
#>    <td style="text-align:left;"> 15% </td>
#>    <td style="text-align:left;"> 12% </td>
#>    <td style="text-align:left;"> 3% </td>
#>   </tr>
#> </tbody>
#> </table>