PortfolioModelTable

This function returns a formatted and styled DT table highlighting the over-weights and under-weights in a portfolio compared to model weights. The table includes color formatting to visually represent the differences.

Usage

PortfolioModelTable(df, mkt_value, port_wgt, model_wgt, active_wgt)

Arguments

df

A data frame containing the input data.

mkt_value

(string) specifying the name of the column containing market values.

port_wgt

(string) specifying the name of the column containing portfolio weights or holdings.

model_wgt

(string) specifying the name of the column containing model allocation weights.

active_wgt

(string)specifying the name of the column containing active weights (difference between portfolio weight and model weight).

Value

A DT object styled for easy visualization.

Examples

# Example usage:
test_df <- data.frame(
  account_code = c('abc', 'xyz', 'test0001'),
  asset_class = c('Equities', 'Bonds', 'Commodities'),
  mkt_value = c(300, 200, 100),
  port_wgt = c(0.5, 0.3, 0.2),
  model_wgt = c(0.4, 0.4, 0.2),
  active_wgt = c(0.1, -0.1, 0.0)
)
PortfolioModelTable(
  df = test_df,
  mkt_value = "mkt_value",
  port_wgt = "port_wgt",
  model_wgt = "model_wgt",
  active_wgt = "active_wgt"
)