ProcessInstrumentMetaID

The ProcessInstrumentMetaID function filters the input dataset based on the specified source names and source types provided via the source_vector and source_type arguments. After filtering, it removes duplicate rows based on the instrument_id column.

Usage

ProcessInstrumentMetaID(data_tbl, source_vector, source_type)

Arguments

data_tbl

A dataset containing Instrument Meta ID information.

source_vector

A character vector specifying the source names to filter.

source_type

A character vector specifying the source types to filter.

Value

A tibble containing the filtered data.

Examples

# Example dataset
data_tbl <- tibble::tibble(
  instrument_id = c(1, 2, 3, 1),
  source_name = c("SourceA", "SourceB", "SourceC", "SourceA"),
  source_type = c("Type1", "Type2", "Type1", "Type1"),
  source_code = c("A123", "B456", "C789", "A123")
)

# Specify source names and types to filter
source_vector <- c("SourceA", "SourceB")
source_type <- c("Type1", "Type2")

result <- ProcessInstrumentMetaID(data_tbl, source_vector, source_type)
print(result)
#> # A tibble: 2 × 3
#>   instrument_id source_type InstrumentCode
#>           <dbl> <chr>       <chr>         
#> 1             1 Type1       A123          
#> 2             2 Type2       B456