SetupMicrosoft365Auth

The SetupMicrosoft365Auth function initializes and creates an authentication token for interacting with Microsoft 365 products. This token is essential when using functionality such as accessing user information within Microsoft 365.

Note: When SetupMicrosoft365Auth is run for the first time, it prompts for authorization via the Authenticator app.

Usage

SetupMicrosoft365Auth(
  tenant = "fairtree.com",
  app_id = "d44a05d5-c6a5-4bbb-82d2-443123722380"
)

Arguments

tenant

A string specifying the Azure AD tenant name or ID.

app_id

A string specifying the client ID (application ID) of your registered Azure AD app.

Value

A Microsoft 365 authentication token object.