AccordionPanel

Creates a collapsible UI panel for displaying grouped bar charts in an interactive application. Each panel section is collapsible and can contain any HTML content, making it ideal for organizing various interactive elements compactly and accessibly.

Usage

AccordionPanel(
  df,
  value_col,
  grouping_col,
  underlying_bars_col,
  chart_title = NULL,
  position = c("auto", "auto", "auto", "auto"),
  width = "auto",
  height = "auto",
  color = "#FF8200",
  underlying_chart_height = 350
)

Arguments

df

A data frame containing the data to be displayed in the panel.

value_col

(string) The column in df that contains the values to be displayed in the chart.

grouping_col

(string) The column that will be used to group the data for each panel.

underlying_bars_col

(string) The column that represents the categories for the bars in the chart.

chart_title

(string, optional) A title for the chart, displayed at the top of the panel.

position

(character vector) A vector of 4 values specifying the position of the panel (top, right, bottom, left). Default is c("auto", "auto", "auto", "auto").

width

(character) The width of the panel. Default is "auto".

height

(character) The height of the panel. Default is "auto".

color

(string or hex) A color value (in hex format or named color). Default is "#FF8200".

underlying_chart_height

(numeric) The height of the individual bar charts in the panel. Default is 350.

Value

An HTML element containing the collapsible panel with grouped bar charts based on ggplot2.

Examples

library(shiny)

# Example data
test_data <- data.frame(
  group_value = rep(c("Group A", "Group B", "Group C"), each = 3),
  category = rep(c("Category 1", "Category 2", "Category 3"), times = 3),
  amount = c(10, 20, 15, 25, 35, 30, 40, 50, 45)
)

AccordionPanel(
  df = test_data,
  value_col = "amount",
  grouping_col = "group_value",
  underlying_bars_col = "category",
  chart_title = "Collapsible Bar Charts"
)
#> <div class="draggable" id="controls" style="top:auto;left:auto;right:auto;bottom:auto;width:auto;height:auto;position:fixed;cursor:move; overflow-y:hidden; max-height: 1500px; opacity: 0.75; z-index: 10000">
#>   <h6>
#>     <strong>Collapsible Bar Charts</strong>
#>   </h6>
#>   <div>
#>     <button data-toggle="collapse" data-target="#boxccb38d2f-54aa-487e-85b5-8763971bba81" class="btn btn-info btn-block">Group A</button>
#>     <div id="boxccb38d2f-54aa-487e-85b5-8763971bba81" class="collapse">
#>       <div id="htmlwidget-ac96cb3ee4656e2e9ec3" style="width:100%;height:350px;" class="plotly html-widget"></div>
#>       <script type="application/json" data-for="htmlwidget-ac96cb3ee4656e2e9ec3">{"x":{"data":[{"orientation":"v","width":[0.5,0.5,0.5],"base":[0,0,0],"x":[3,1,2],"y":[10,20,15],"text":["reorder(category, -amount): Category 1<br />amount: 10","reorder(category, -amount): Category 2<br />amount: 20","reorder(category, -amount): Category 3<br />amount: 15"],"type":"bar","textposition":"none","marker":{"autocolorscale":false,"color":"rgba(255,130,0,1)","line":{"width":1.8897637795275593,"color":"transparent"}},"showlegend":false,"xaxis":"x","yaxis":"y","hoverinfo":"text","frame":null}],"layout":{"margin":{"t":35.128268991282695,"r":13.283520132835205,"b":51.514008472636732,"l":53.33333333333335},"plot_bgcolor":"rgba(255,255,255,1)","paper_bgcolor":"rgba(255,255,255,1)","font":{"color":"rgba(0,0,0,1)","family":"","size":14.611872146118724},"xaxis":{"domain":[0,1],"automargin":true,"type":"linear","autorange":false,"range":[0.40000000000000002,3.6000000000000001],"tickmode":"array","ticktext":["Category 2","Category 3","Category 1"],"tickvals":[1,2,3],"categoryorder":"array","categoryarray":["Category 2","Category 3","Category 1"],"nticks":null,"ticks":"","tickcolor":null,"ticklen":3.6529680365296811,"tickwidth":0,"showticklabels":true,"tickfont":{"color":"rgba(77,77,77,1)","family":"","size":9.2984640929846432},"tickangle":-45,"showline":false,"linecolor":null,"linewidth":0,"showgrid":false,"gridcolor":null,"gridwidth":0,"zeroline":false,"anchor":"y","title":{"text":"","font":{"color":null,"family":null,"size":0}},"hoverformat":".2f"},"yaxis":{"domain":[0,1],"automargin":true,"type":"linear","autorange":false,"range":[-1,21],"tickmode":"array","ticktext":["0%","500%","1 000%","1 500%","2 000%"],"tickvals":[0,5,10,15,20],"categoryorder":"array","categoryarray":["0%","500%","1 000%","1 500%","2 000%"],"nticks":null,"ticks":"outside","tickcolor":"rgba(51,51,51,1)","ticklen":3.6529680365296811,"tickwidth":0.66417600664176002,"showticklabels":true,"tickfont":{"color":"rgba(77,77,77,1)","family":"","size":11.68949771689498},"tickangle":-0,"showline":false,"linecolor":null,"linewidth":0,"showgrid":false,"gridcolor":null,"gridwidth":0,"zeroline":false,"anchor":"x","title":{"text":"","font":{"color":"rgba(0,0,0,1)","family":"","size":14.611872146118724}},"hoverformat":".2f"},"shapes":[{"type":"rect","fillcolor":null,"line":{"color":null,"width":0,"linetype":[]},"yref":"paper","xref":"paper","x0":0,"x1":1,"y0":0,"y1":1}],"showlegend":false,"legend":{"bgcolor":"rgba(255,255,255,1)","bordercolor":"transparent","borderwidth":1.8897637795275593,"font":{"color":"rgba(0,0,0,1)","family":"","size":11.68949771689498}},"hovermode":"closest","height":350,"barmode":"relative"},"config":{"doubleClick":"reset","modeBarButtonsToAdd":["hoverclosest","hovercompare"],"showSendToCloud":false},"source":"A","attrs":{"57f4167a2f56":{"x":{},"y":{},"type":"bar"}},"cur_data":"57f4167a2f56","visdat":{"57f4167a2f56":["function (y) ","x"]},"highlight":{"on":"plotly_click","persistent":false,"dynamic":false,"selectize":false,"opacityDim":0.20000000000000001,"selected":{"opacity":1},"debounce":0},"shinyEvents":["plotly_hover","plotly_click","plotly_selected","plotly_relayout","plotly_brushed","plotly_brushing","plotly_clickannotation","plotly_doubleclick","plotly_deselect","plotly_afterplot","plotly_sunburstclick"],"base_url":"https://plot.ly"},"evals":[],"jsHooks":[]}</script>
#>     </div>
#>   </div>
#>   <div>
#>     <button data-toggle="collapse" data-target="#box2caa2516-b538-4de3-9cdd-78d8133effad" class="btn btn-info btn-block">Group B</button>
#>     <div id="box2caa2516-b538-4de3-9cdd-78d8133effad" class="collapse">
#>       <div id="htmlwidget-e5c8c404fe174e4c81bd" style="width:100%;height:350px;" class="plotly html-widget"></div>
#>       <script type="application/json" data-for="htmlwidget-e5c8c404fe174e4c81bd">{"x":{"data":[{"orientation":"v","width":[0.5,0.5,0.5],"base":[0,0,0],"x":[3,1,2],"y":[25,35,30],"text":["reorder(category, -amount): Category 1<br />amount: 25","reorder(category, -amount): Category 2<br />amount: 35","reorder(category, -amount): Category 3<br />amount: 30"],"type":"bar","textposition":"none","marker":{"autocolorscale":false,"color":"rgba(255,130,0,1)","line":{"width":1.8897637795275593,"color":"transparent"}},"showlegend":false,"xaxis":"x","yaxis":"y","hoverinfo":"text","frame":null}],"layout":{"margin":{"t":35.128268991282695,"r":13.283520132835205,"b":51.514008472636732,"l":53.33333333333335},"plot_bgcolor":"rgba(255,255,255,1)","paper_bgcolor":"rgba(255,255,255,1)","font":{"color":"rgba(0,0,0,1)","family":"","size":14.611872146118724},"xaxis":{"domain":[0,1],"automargin":true,"type":"linear","autorange":false,"range":[0.40000000000000002,3.6000000000000001],"tickmode":"array","ticktext":["Category 2","Category 3","Category 1"],"tickvals":[1,2,3],"categoryorder":"array","categoryarray":["Category 2","Category 3","Category 1"],"nticks":null,"ticks":"","tickcolor":null,"ticklen":3.6529680365296811,"tickwidth":0,"showticklabels":true,"tickfont":{"color":"rgba(77,77,77,1)","family":"","size":9.2984640929846432},"tickangle":-45,"showline":false,"linecolor":null,"linewidth":0,"showgrid":false,"gridcolor":null,"gridwidth":0,"zeroline":false,"anchor":"y","title":{"text":"","font":{"color":null,"family":null,"size":0}},"hoverformat":".2f"},"yaxis":{"domain":[0,1],"automargin":true,"type":"linear","autorange":false,"range":[-1.75,36.75],"tickmode":"array","ticktext":["0%","1 000%","2 000%","3 000%"],"tickvals":[0,10,20,30],"categoryorder":"array","categoryarray":["0%","1 000%","2 000%","3 000%"],"nticks":null,"ticks":"outside","tickcolor":"rgba(51,51,51,1)","ticklen":3.6529680365296811,"tickwidth":0.66417600664176002,"showticklabels":true,"tickfont":{"color":"rgba(77,77,77,1)","family":"","size":11.68949771689498},"tickangle":-0,"showline":false,"linecolor":null,"linewidth":0,"showgrid":false,"gridcolor":null,"gridwidth":0,"zeroline":false,"anchor":"x","title":{"text":"","font":{"color":"rgba(0,0,0,1)","family":"","size":14.611872146118724}},"hoverformat":".2f"},"shapes":[{"type":"rect","fillcolor":null,"line":{"color":null,"width":0,"linetype":[]},"yref":"paper","xref":"paper","x0":0,"x1":1,"y0":0,"y1":1}],"showlegend":false,"legend":{"bgcolor":"rgba(255,255,255,1)","bordercolor":"transparent","borderwidth":1.8897637795275593,"font":{"color":"rgba(0,0,0,1)","family":"","size":11.68949771689498}},"hovermode":"closest","height":350,"barmode":"relative"},"config":{"doubleClick":"reset","modeBarButtonsToAdd":["hoverclosest","hovercompare"],"showSendToCloud":false},"source":"A","attrs":{"57f4465f5d07":{"x":{},"y":{},"type":"bar"}},"cur_data":"57f4465f5d07","visdat":{"57f4465f5d07":["function (y) ","x"]},"highlight":{"on":"plotly_click","persistent":false,"dynamic":false,"selectize":false,"opacityDim":0.20000000000000001,"selected":{"opacity":1},"debounce":0},"shinyEvents":["plotly_hover","plotly_click","plotly_selected","plotly_relayout","plotly_brushed","plotly_brushing","plotly_clickannotation","plotly_doubleclick","plotly_deselect","plotly_afterplot","plotly_sunburstclick"],"base_url":"https://plot.ly"},"evals":[],"jsHooks":[]}</script>
#>     </div>
#>   </div>
#>   <div>
#>     <button data-toggle="collapse" data-target="#box2229ed7f-f103-486e-b11d-c7c0761cbb75" class="btn btn-info btn-block">Group C</button>
#>     <div id="box2229ed7f-f103-486e-b11d-c7c0761cbb75" class="collapse">
#>       <div id="htmlwidget-36aa3d2a04d42bbc2145" style="width:100%;height:350px;" class="plotly html-widget"></div>
#>       <script type="application/json" data-for="htmlwidget-36aa3d2a04d42bbc2145">{"x":{"data":[{"orientation":"v","width":[0.5,0.5,0.5],"base":[0,0,0],"x":[3,1,2],"y":[40,50,45],"text":["reorder(category, -amount): Category 1<br />amount: 40","reorder(category, -amount): Category 2<br />amount: 50","reorder(category, -amount): Category 3<br />amount: 45"],"type":"bar","textposition":"none","marker":{"autocolorscale":false,"color":"rgba(255,130,0,1)","line":{"width":1.8897637795275593,"color":"transparent"}},"showlegend":false,"xaxis":"x","yaxis":"y","hoverinfo":"text","frame":null}],"layout":{"margin":{"t":35.128268991282695,"r":13.283520132835205,"b":51.514008472636732,"l":53.33333333333335},"plot_bgcolor":"rgba(255,255,255,1)","paper_bgcolor":"rgba(255,255,255,1)","font":{"color":"rgba(0,0,0,1)","family":"","size":14.611872146118724},"xaxis":{"domain":[0,1],"automargin":true,"type":"linear","autorange":false,"range":[0.40000000000000002,3.6000000000000001],"tickmode":"array","ticktext":["Category 2","Category 3","Category 1"],"tickvals":[1,2,3],"categoryorder":"array","categoryarray":["Category 2","Category 3","Category 1"],"nticks":null,"ticks":"","tickcolor":null,"ticklen":3.6529680365296811,"tickwidth":0,"showticklabels":true,"tickfont":{"color":"rgba(77,77,77,1)","family":"","size":9.2984640929846432},"tickangle":-45,"showline":false,"linecolor":null,"linewidth":0,"showgrid":false,"gridcolor":null,"gridwidth":0,"zeroline":false,"anchor":"y","title":{"text":"","font":{"color":null,"family":null,"size":0}},"hoverformat":".2f"},"yaxis":{"domain":[0,1],"automargin":true,"type":"linear","autorange":false,"range":[-2.5,52.5],"tickmode":"array","ticktext":["0%","1 000%","2 000%","3 000%","4 000%","5 000%"],"tickvals":[0,10,20,30,40,50],"categoryorder":"array","categoryarray":["0%","1 000%","2 000%","3 000%","4 000%","5 000%"],"nticks":null,"ticks":"outside","tickcolor":"rgba(51,51,51,1)","ticklen":3.6529680365296811,"tickwidth":0.66417600664176002,"showticklabels":true,"tickfont":{"color":"rgba(77,77,77,1)","family":"","size":11.68949771689498},"tickangle":-0,"showline":false,"linecolor":null,"linewidth":0,"showgrid":false,"gridcolor":null,"gridwidth":0,"zeroline":false,"anchor":"x","title":{"text":"","font":{"color":"rgba(0,0,0,1)","family":"","size":14.611872146118724}},"hoverformat":".2f"},"shapes":[{"type":"rect","fillcolor":null,"line":{"color":null,"width":0,"linetype":[]},"yref":"paper","xref":"paper","x0":0,"x1":1,"y0":0,"y1":1}],"showlegend":false,"legend":{"bgcolor":"rgba(255,255,255,1)","bordercolor":"transparent","borderwidth":1.8897637795275593,"font":{"color":"rgba(0,0,0,1)","family":"","size":11.68949771689498}},"hovermode":"closest","height":350,"barmode":"relative"},"config":{"doubleClick":"reset","modeBarButtonsToAdd":["hoverclosest","hovercompare"],"showSendToCloud":false},"source":"A","attrs":{"57f476e377d2":{"x":{},"y":{},"type":"bar"}},"cur_data":"57f476e377d2","visdat":{"57f476e377d2":["function (y) ","x"]},"highlight":{"on":"plotly_click","persistent":false,"dynamic":false,"selectize":false,"opacityDim":0.20000000000000001,"selected":{"opacity":1},"debounce":0},"shinyEvents":["plotly_hover","plotly_click","plotly_selected","plotly_relayout","plotly_brushed","plotly_brushing","plotly_clickannotation","plotly_doubleclick","plotly_deselect","plotly_afterplot","plotly_sunburstclick"],"base_url":"https://plot.ly"},"evals":[],"jsHooks":[]}</script>
#>     </div>
#>   </div>
#> </div>
#> <script>$(".draggable").draggable({cancel: ".shiny-input-container,.html-widget,input,textarea,button,select,option"});</script>